feat: implement persistent moderation logging and strike system
